This commit was manufactured by cvs2svn to create tag 'RELENG_3_0_0_RELEASE'.

+++ include/protocols/rwhod.h
@@ -42,7 +42,7 @@
 struct	outmp {
 	char	out_line[8];		/* tty name */
 	char	out_name[8];		/* user id */
-	int32_t	out_time;		/* time on */
+	long	out_time;		/* time on */
 };
 
 struct	whod {
